Fix python TODO comments
